+++ src/widgets_json.h
@@ -5,6 +5,6 @@
 #include "widgets.h"
 
 int deserialise_widgets_file(const char* file_path, view_context_t* ctx);
-int deserialise_json(const char* json_string, const int len, view_context_t* ctx);
+int deserialise_json(const char* json_string, const size_t len, view_context_t* ctx);
 
 #endif  // __jl_widgets_json_h_
